Move filter_hdulist_by_shape to deprecated functions
