+++ glue/core/data_factories/fits.py
@@ -19,66 +19,6 @@ def is_fits(filename):
         return False
 
 
-def filter_hdulist_by_shape(hdulist, use_hdu='all'):
-    """
-    Remove empty HDUs, and ensure that all HDUs can be
-    packed into a single Data object (ie have the same shape)
-
-    Parameters
-    ----------
-    use_hdu : 'all' or list of integers (optional)
-        Which HDUs to use
-
-    Returns
-    -------
-    a new HDUList
-    """
-    from ...external.astro import fits
-
-    # If only a subset are requested, extract those
-    if use_hdu != 'all':
-        hdulist = [hdulist[hdu] for hdu in use_hdu]
-
-    # Now only keep HDUs that are not tables or empty.
-    valid_hdus = []
-    for hdu in hdulist:
-        if (isinstance(hdu, fits.PrimaryHDU) or
-                isinstance(hdu, fits.ImageHDU)) and \
-                hdu.data is not None:
-            valid_hdus.append(hdu)
-
-    # Check that dimensions of all HDU are the same
-    # Allow for HDU's that have no data.
-    reference_shape = valid_hdus[0].data.shape
-    for hdu in valid_hdus:
-        if hdu.data.shape != reference_shape:
-            raise Exception("HDUs are not all the same dimensions")
-
-    return valid_hdus
-
-
-def extract_data_fits(filename, use_hdu='all'):
-    '''
-    Extract non-tabular HDUs from a FITS file. If `use_hdu` is 'all', then
-    all non-tabular HDUs are extracted, otherwise only the ones specified
-    by `use_hdu` are extracted (`use_hdu` should then contain a list of
-    integers). If the requested HDUs do not have the same dimensions, an
-    Exception is raised.
-    '''
-    from ...external.astro import fits
-
-    # Read in all HDUs
-    hdulist = fits.open(filename, ignore_blank=True)
-    hdulist = filter_hdulist_by_shape(hdulist)
-
-    # Extract data
-    arrays = {}
-    for hdu in hdulist:
-        arrays[hdu.name] = hdu.data
-
-    return arrays
-
-
 @data_factory(
     label='FITS file',
     identifier=is_fits,
@@ -172,6 +112,7 @@ def fits_reader(source, auto_merge=False, exclude_exts=None):
 
 
 # Utilities
+
 def is_image_hdu(hdu):
     from astropy.io.fits.hdu import PrimaryHDU, ImageHDU
     return isinstance(hdu, (PrimaryHDU, ImageHDU))
